requirements.txt
requests>=2.31.0
pytest>=7.4.0
PyYAML>=6.0.1
allure-pytest>=2.13.2
openpyxl>=3.1.2
pytest-dependency>=0.5.1
tenacity>=8.2.3
pycryptodome>=3.19.0
python-dotenv>=26.0.1
